🚚 Move express app out of server.js
